Add: views.py for Forgot Password
In this push I have added the views code for forgot password where when a user requests a password reset, an OTP is generated, stored in the session and sent to the user via email. The user is then prompted to enter the OTP for verification. If the OTP is correct user is directed to reset their password.