Update: Show Password & Add Condition for not Matching New and Current Password
In this commit, I have added the functionality to show the password using JS and a checkbox button to show and hide the password. Here, I have added condition for button where until the data is updated the message cannot be pop up. Also, added a condition to check if the new password and old password matches or not.